first trial

+++ dataloader.py
@@ -99,7 +99,7 @@ def convert_batch_to_tensors(batch):
     # labels = np.reshape(labels,(labels.shape[0],1))
     print(labels.shape)
     images = torch.Tensor(images)
-    labels = torch.LongTensor(labels)
+    labels = torch.LongTensor(labels).float()
     
     return images,labels
